@@ -15,23 +15,24 @@
   "author": "Nicholas Frai",
   "license": "MIT",
   "dependencies": {
-    "bcrypt": "^3.0.6",
-    "cookie-parser": "^1.4.4",
+    "bcrypt": "^3.0.8",
+    "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
+    "cryptr": "^6.0.1",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
-    "knex": "^0.20.3",
-    "mysql": "^2.17.1",
+    "knex": "^0.20.11",
+    "mysql": "^2.18.1",
     "mysql2": "^1.6.5"
   },
   "devDependencies": {
     "@types/bcrypt": "^3.0.0",
     "@types/event-stream": "^3.3.34",
-    "@types/express": "^4.17.0",
-    "@types/jsonwebtoken": "^8.3.2",
-    "@types/mysql": "^2.15.6",
-    "nodemon": "^1.19.4",
-    "typescript": "^3.5.3"
+    "@types/express": "^4.17.3",
+    "@types/jsonwebtoken": "^8.3.8",
+    "@types/mysql": "^2.15.9",
+    "nodemon": "^2.0.2",
+    "typescript": "^3.8.3"
   }
 }
