@@ -1,13 +1,14 @@
 import { Express, Request, Response, NextFunction } from 'express-serve-static-core';
-import { AuriaCoreSystem } from './system/AuriaCore/AuriaCoreSystem';
 import { System } from './kernel/System';
 import { SystemRequest } from './kernel/http/request/SystemRequest';
 import { SystemUnavaliable } from './kernel/exceptions/kernel/SystemUnavaliable';
 import { RequestStack } from './kernel/RequestStack';
 import { ServerRequest, ServerRequestFactory } from './kernel/http/request/ServerRequest';
-import { response } from 'express';
 import { AuriaException } from './kernel/exceptions/AuriaException';
-import e = require('express');
+import { AuriaServerBootInfo } from './server/AuriaServerBootInfo';
+import { AuriaCoreSystem } from './system/AuriaCore/AuriaCoreSystem';
+import { AuriaSystem } from './default/AuriaSystem';
+import { ServerResponse } from 'aurialib2';
 
 export type AuriaServerStatus = "online" | "offline" | "maintenance";
 
@@ -46,7 +47,9 @@ export class AuriaServer {
     /**
      * This server session
      */
-    private serverSessionId: number;
+    private serverInstanceId: number;
+
+    protected auriaBootInfo: AuriaServerBootInfo;
 
 
     /**
@@ -63,63 +66,96 @@ export class AuriaServer {
             try {
 
                 let serverReq: ServerRequest = ServerRequestFactory.promote(req);
+                let stack: RequestStack = RequestStack.digestRequest(req);
 
-                let stack: RequestStack = RequestStack.digestURL(req.url);
+                if (stack.system() == "") {
+                    this.answerServerStatus(res);
+                    return;
+                }
 
                 if (!this.systems.has(stack.system())) {
-                    throw new SystemUnavaliable("[SystemRequest] The requested system is not avaliable on this server!");
+                    throw new SystemUnavaliable("[Server] The requested system is not avaliable on this server! - " + stack.system());
                 }
 
                 // # ELSE, system exists in server
                 let system = this.systems.get(stack.system())!;
                 let systemRequest: SystemRequest = system.promoteToSystemRequest(serverReq, stack);
-                let systemResponse = system.handleRequest(systemRequest, res, next);
+                
+                // Sending the data through the Response object is a 'System' Responsability!
+                let systemResponse = await system.handleRequest(systemRequest, res, next);
 
-                if (systemResponse instanceof Promise) {
-                    systemResponse.then((ans: any) => {
-                        console.log("[Server] Respons to request: ", ans);
-                        res.send(ans);
-                    }).catch((err) => {
-                        let exc = err as AuriaException;
-                        console.error("[Server] Failed to process request!", exc);
-                        res.status(400);
-                        res.send({
-                            code: exc.getCode(),
-                            message: exc.getMessage()
-                        });
-                    });
-                } else {
-
-                }
                 console.log("System Response to request: ", systemResponse);
             }
             catch (ex) {
-
+                let exc = ex as AuriaException;
+                this.handleRequestException(exc, res);
             }
         };
 
+    private answerServerStatus(response: Response) {
+        
+        let status: ServerResponse = {
+            digest: "ok",
+            error :"",
+            exitCode: "SERVER.STATUS",
+            response: {
+                server_status: this.serverStatus
+            }
+        };
+
+        response.send(status);
+    }
+
+    private handleRequestException(exception: AuriaException, response: Response) {
+
+        console.error("[Server] Failed to proccess request!", exception);
+
+        response.status(exception.getHttpCode());
+
+        let sendArgs: any = {
+            digest: "error",
+            code: exception.getCode(),
+            message: exception.getMessage()
+        };
+
+        if (exception.getExtraArgs().length > 0) {
+            sendArgs.extra = exception.getExtraArgs();
+        }
+
+        response.send(sendArgs);
+    }
 
     constructor(app: Express) {
 
         console.log("\n[Auria Server] Initializing a new Auria server!");
 
         this.app = app;
-        this.initializeExpressApp();
-
         this.systems = new Map();
-        this.serverSessionId = Math.round(Math.random() * 10000000);
+        this.serverInstanceId = Math.round(Math.random() * 10000000);
 
-        console.log("[Auria Server] Initializing Systems...");
+        this.auriaBootInfo = {
+            boot_server_random_id: this.serverInstanceId,
+            boot_timestamp: Date.now()
+        };
+
+        this.initializeExpressApp();
 
         this.addSystem(
             new AuriaCoreSystem(),
-            // new Este()
+            new AuriaSystem("test")
         );
-
-        console.log("[Auria Server] Server Instance Token: " + this.serverSessionId);
 
     }
 
+    /**
+     * Load Express modules
+     * --------------------
+     * Common to ALL SYSTEMS under this server!
+     * > BodyParser: URL Encoded + JSON
+     * > CookieParser
+     * > Disable X-Powered-By Header (security trough obscurity?)
+     * > CORS (?)
+     */
     private initializeExpressApp() {
 
         var bodyParser = require('body-parser');
@@ -140,10 +176,8 @@ export class AuriaServer {
      */
     public addSystem(...system: System[]): AuriaServer {
         system.forEach((sys) => {
-            console.log("[Auria Server] Adding system: ", sys.name);
             this.systems.set(sys.name, sys);
         });
-
         return this;
     }
 
@@ -157,3 +191,5 @@ export class AuriaServer {
         return this;
     }
 }
+
+export const AURIA_LOG_ROOT = __dirname + "/logs";
