@@ -1,5 +1,5 @@
 import { Response, NextFunction } from "express-serve-static-core";
-
+import { EventEmitter } from 'events';
 import { SystemRequest, SystemRequestFactory } from "./http/request/SystemRequest";
 import { SystemAuthenticator } from "./security/auth/SystemAuthenticator";
 import { ModuleManager } from "./module/ModuleManager";
@@ -8,11 +8,10 @@ import { ModuleRequestFactory } from "./http/request/ModuleRequest";
 import { RequestStack } from "./RequestStack";
 
 // Default system modules
-import { AuthModule } from "./module/AuthModule/AuthModule";
 import { SystemModule } from "./module/SystemModule/SystemModule";
 
 import { SystemUser } from "./security/SystemUser";
-import { DataSteward } from "aurialib2";
+import { DataSteward, Authenticator, ServerResponse } from "aurialib2";
 import knex from 'knex';
 
 // Params config import
@@ -20,12 +19,21 @@ import { Auria_ENV } from "../AuriaServer";
 // Exceptions
 import { ModuleUnavaliable } from "./exceptions/kernel/ModuleUnavaliable";
 import { ServerRequest } from "./http/request/ServerRequest";
-import { LoginRequest } from "./module/SystemModule/requests/LoginRequest";
+import { AccessPolicyEnforcer } from "./security/access/AccessPolicyEnforcer";
+import { UnauthorizedAccess } from "./exceptions/kernel/UnauthorizedAccess";
+import Knex from "knex";
+import { UserNotLoggedIn } from "./exceptions/kernel/UserNotLoggedIn";
+import { AccessRuleFactory } from "./security/access/AccessRuleFactory";
+import { AuriaException } from "./exceptions/AuriaException";
 
 export const DEFAULT_LANG = "en";
 export const DEFAULT_LANG_VARIATION = "us";
 
-export abstract class System {
+export abstract class System extends EventEmitter {
+
+    public static EVENT_SYSTEM_MODULE_ADDED = "SystemModuleAdded";
+
+    public static EVENT_SYSTEM_RESOURCE_MODIFIED = "SystemResourceModified";
 
     /**
      * System name
@@ -45,7 +53,6 @@ export abstract class System {
      */
     protected systemVersion: number;
 
-
     /**
      * Module manager
      * 
@@ -57,22 +64,49 @@ export abstract class System {
     /**
      * System connection
      * ------------------
-     * 
      */
-    protected connection: knex;
+    protected connection: Knex;
+
+    /**
+     * System Authenticator
+     * ---------------------
+     * 
+     * Shall parse authentication from the HTTP requests and validate them
+     * Generating credentials and/or JWTokens is the responsability of the 
+     * System Module > Login Listener!
+     * 
+     * Authenticator only validates the HTTP credentials and associates them to a
+     * user!
+     */
+    protected authenticator: Authenticator;
 
     /**
      * Data Steward
      * -------------
      * 
+     * DataSteward is the gateway to accessing and modifying information
+     * on the database!
      */
     protected dataSteward: DataSteward;
+
+    /**
+     * Access Policy Enforcer
+     * ----------------------
+     * 
+     * Checks if the user have permission to access the requested action!
+     * 
+     * Each "<ListenerAction>" have its own "<AccessPolicy>" determined by their 
+     * <ModuleListener> AccessPolicyEnforcer Traverse through each module gathering
+     * all of this Access Policies and applies them in each Request!
+     */
+    protected accessPolicyEnforcer: AccessPolicyEnforcer;
 
     /**
      * Translations
      * ------------
      * 
      * Hold all the loaded translations from this server
+     * (Should be here?)
      */
     protected loadedTranslations: {
         [langVariation: string]: any
@@ -82,50 +116,51 @@ export abstract class System {
      * [Factory] SystemRequest
      * ------------------------
      * 
-     * Factory to produce the expected SystemRequest
+     * Factory to produce a <SystemRequest> Object
+     * A SystemRequest is an extension of the express <Request> Object
+     * 
+     * The factory shall add functions and variables that is AVALIABLE TO ALL
+     * LEVES OF THE REQUEST ( System > Module > Listener > Action );
      * 
      */
     protected systemRequestFactory: SystemRequestFactory;
 
     constructor(name: string) {
+        super();
+
+        console.info("[System] Creating new system: ", name);
         this.name = name;
+
         this.systemRequestFactory = new SystemRequestFactory();
-
-        console.log("[System] Creating new system: ", name);
-
         this.moduleManager = new ModuleManager(this);
+        this.accessPolicyEnforcer = new AccessPolicyEnforcer(this);
+        this.users = new Map();
+        this.accessPolicyEnforcer.setAccessRuleFactory(this.getAccessRuleFactory());
 
         // If ENV == "development", systemversion does not change!
-        if (Auria_ENV == "development")
-            this.systemVersion = 1;
-        else
-            this.systemVersion = Math.round(Math.random() * 1000000);
-
-        this.users = new Map();
+        this.systemVersion = Auria_ENV == "development" ?
+            1 : Math.round(Math.random() * 1000000);
 
         console.log("[System] Initializing modules from system ", name);
-
+       
         this.addModule(
-            // # - Authentication Module
-            new AuthModule(this),
-
-            // # - System related functions Module
+            // # - System related functions
             new SystemModule(this)
         );
     }
 
     /**
-     * Build all modules from this system
-     * 
-     * This function is called one time at server startup!
-     * 
-     */
-    protected abstract buildSystemModules(): Map<string, Module>;
-
-    /**
      * Public access to this system modules instances
      */
-    public abstract getSystemModules(): Map<string, Module>;
+    public getSystemModules(): Map<string, Module> {
+        let modules: Map<string, Module> = new Map();
+
+        this.moduleManager.getAllModules().forEach((mod) => {
+            modules.set(mod.name, mod);
+        });
+
+        return modules;
+    }
 
     /**
      * Build access to this system auria connection
@@ -138,22 +173,38 @@ export abstract class System {
     public abstract getSystemConnection(): knex;
 
     /**
-     * 
+     * Public access to this system authenticator
      */
     public abstract getAuthenticator(): SystemAuthenticator;
 
-
-    public loginUser(user : SystemUser, request : LoginRequest) : System {
+    /**
+     * Login User 
+     * -----------
+     * 
+     * Register an user as "logged in" in this system!
+     * 
+     * -- Alias to 'addUser' --
+     * @param user 
+     * @param request 
+     */
+    public loginUser(user: SystemUser, request: SystemRequest): System {
         return this.addUser(user, request);
     }
 
     /**
+     * Add User
+     * ---------
      * 
-     * @param user 
-     * @param request 
+     * Adds an user to this system, the user will be considered "logged in" and
+     * future requests with a valid authentication attached to this user will be
+     * accepted 
+     * 
+     * @param user To user to be added 
+     * @param request The request to this user be logged in
+     * 
      */
-    public addUser(user: SystemUser, request : LoginRequest): System {
-        
+    public addUser(user: SystemUser, request: SystemRequest): System {
+
         user.startSession(request);
 
         this.users.set(user.getUsername(), user);
@@ -168,17 +219,18 @@ export abstract class System {
         return this.moduleManager.hasModule(moduleName);
     }
 
-    public addModule(...module: Module[]) {
-        module.forEach((mod) => {
-            let translations = mod.getTranslations();
-
-            for (var lang in translations) {
-                if (translations.hasOwnProperty(lang)) {
-                }
-            }
-
+    /**
+     * Add Module
+     * ------------
+     * Add a Module to this system!
+     * 
+     * @param modules 
+     */
+    public addModule(...modules: Module[]) {
+        modules.forEach((mod) => {
             this.moduleManager.addModule(mod);
-        })
+            this.emit(System.EVENT_SYSTEM_MODULE_ADDED, mod);
+        });
     }
 
     public getModule(moduleName: string) {
@@ -219,8 +271,22 @@ export abstract class System {
      * @param next Express **NextFunction**
      */
     public async handleRequest(request: SystemRequest, response: Response, next: NextFunction) {
+        let user: SystemUser;
 
-        let user: SystemUser = await this.getAuthenticator().authenticateRequest(request);
+        try {
+            user = await this.authenticateRequest(request);
+        } catch (err) {
+            console.log("FAILED TO AUTHENTICATE: ", err, err instanceof UserNotLoggedIn);
+            user = new SystemUser(this, SystemUser.GUEST_USERNAME);
+        }
+
+        console.log("[Request] Checking user permission!");
+        let authorized = await this.accessPolicyEnforcer.authorize(user, request);
+
+        if (!authorized) {
+            throw new UnauthorizedAccess("The current user is not elegible to access this resource!");
+        }
+
         let requestedModule = request.getRequestStack().module();
 
         if (!this.moduleManager.hasModule(requestedModule)) {
@@ -230,7 +296,96 @@ export abstract class System {
         let module: Module = this.moduleManager.getModule(requestedModule)!;
         let moduleRequest = ModuleRequestFactory.make(request, user, module);
 
-        return module.handleRequest(moduleRequest, response);
+        let requestResponse = module.handleRequest(moduleRequest, response);
+        let resp = await this.systemResponseFactory(requestResponse, response);
+        response.send(resp);
+        return resp;
+
+    }
+
+    /**
+     * System Response Factory
+     * ------------------------
+     * 
+     * Sanitize the action output response
+     * Usefull to create a standart response to request
+     * 
+     * @param data 
+     * @param response 
+     */
+    public async systemResponseFactory(data: any, response: Response): Promise<ServerResponse> {
+
+        /**
+         * HTTP Response should be handled by the system itself and not the server!
+         * 
+         * Server is just a container for systems, an entrypoint
+         */
+        if (data instanceof Promise) {
+            return data
+                .then((ans: any) => {
+                    console.log("[Server] Response to request: ", ans);
+                    let srvResponse: ServerResponse = {
+                        digest: "ok",
+                        error: "",
+                        exitCode: "SYSTEM.REQUEST.SUCCESS",
+                        response: ans
+                    };
+                    return srvResponse;
+                }).catch((err) => {
+                    let exc = err as AuriaException;
+                    return this.handleRequestException(exc, response);
+                });
+        } else {
+            console.log("[Server] Response to request: ", data);
+            let srvResponse: ServerResponse = {
+                digest: "ok",
+                error: "",
+                exitCode: "SYSTEM.REQUEST.SUCCESS",
+                response: data
+            };
+            return srvResponse;
+        }
+    }
+
+    /**
+     * Handle Request Exception
+     * -------------------------
+     * 
+     * How an Request Exception should be treated!
+     * 
+     * Notice that are 2 types of "Exceptions"! 
+     * One occurs from flaws in the code (dividing by zero, accessing null object) 
+     * others are exceptions in the flow of the request caused by lack of permission
+     * or trying to perform unauthorized changes in the system!
+     *  
+     * @param exception 
+     * @param response 
+     */
+    protected handleRequestException(exception: AuriaException, response: Response): ServerResponse {
+
+        console.error("[Server] Failed to proccess request!", exception);
+
+        response.status(exception.getHttpCode());
+
+        let sendArgs: ServerResponse = {
+            digest: "error",
+            exitCode: exception.getCode(),
+            error: exception.getCode(),
+            response: exception.getMessage()
+        };
+
+        return sendArgs;
+    }
+
+    /**
+     * Authenticate Request
+     * ---------------------
+     * 
+     * Validate the credentials in the request
+     * @param request 
+     */
+    public async authenticateRequest(request: SystemRequest) {
+        return this.getAuthenticator().authenticateRequest(request);
     }
 
     /**
@@ -244,5 +399,14 @@ export abstract class System {
     public promoteToSystemRequest(request: ServerRequest, stack: RequestStack): SystemRequest {
         return this.systemRequestFactory.make(request, this, stack);
     }
+
+    /**
+     * Access Rule Factory
+     * --------------------
+     * 
+     * An Access Rule Factory will transform the AccessRules exposed by the ModuleListeners
+     * in actual AccessRule Objects! 
+     */
+    protected abstract getAccessRuleFactory(): AccessRuleFactory;
 
 }
