@@ -1,12 +1,10 @@
-import { ModuleListener, ListenerActionsDefinition, ListenerAction } from "../../../../../kernel/module/ModuleListener";
-import { AuriaRequest } from "../../../../../kernel/http/AuriaRequest";
-import { AuriaResponse } from "../../../../../kernel/http/AuriaResponse";
-import { AuriaMiddleware } from "../../../../../kernel/http/AuriaMiddleware";
+import { ModuleListener } from "../../../../../kernel/module/ModuleListener";
 import { Table } from "../../../../../kernel/database/structure/table/Table";
 import { AuriaConnection } from "../../../../../kernel/database/connection/AuriaConnection";
 import { SystemUser } from "../../../../../kernel/security/SystemUser";
 import { AuriaArchitect } from "../AuriaArchitect";
 import { DatabaseSychronizer } from "../databaseManipulation/DatabaseSynchronizer";
+import { ListenerActionsMetadata, ListenerAction } from "../../../../../kernel/module/ListenerAction";
 
 export class TableManagerListener extends ModuleListener {
 
@@ -19,11 +17,8 @@ export class TableManagerListener extends ModuleListener {
 
     }
 
-    public getRequiredRequestHandlers(): AuriaMiddleware[] {
-        return [];
-    }
 
-    public getExposedActionsDefinition(): ListenerActionsDefinition {
+    public getExposedActionsMetadata(): ListenerActionsMetadata {
         return {
             "list": {},
             "databaseSync": {},
@@ -104,58 +99,58 @@ export class TableManagerListener extends ModuleListener {
 
     private async createColumnFromComparisson(user: SystemUser, table: Table, comparisson: ColumnComparisson) {
         throw new Error("Not implemented yet");
-/*
-        try {
-
-            let [colTable, txtResTable] = await Promise.all([
-                this.module.getTable(user, "Auria.Collumn"),
-                this.module.getTable(user, "Auria.TextResource")]
-            );
-
-            let colName = this.toCamelCase(comparisson.outOfSync!.field);
-            let colTitleKey = "@{Auria.Columns.Title." + table.getName() + "." + colName + "}";
-            let colDescriptionKey = "@{Auria.Columns.Description." + table.getName() + "." + colName + "}";
-            let dataType = this.rawTypeToDataType(comparisson.outOfSync!.type!);
-
-            // # - Create and save column
-            let col = new RowModel(colTable!);
-            col.setAttribute({
-                name: colName,
-                column: comparisson.outOfSync!.field,
-                table_name: table.table,
-                title: colTitleKey,
-                description: colDescriptionKey,
-                table_type: "physic",
-                data_type: dataType,
-                attributes: comparisson.outOfSync!.key,
-                required: 1,
-                allow_modification: 1,
-                active: 1
-            });
-            col.save(user);
-
-            // # - Create and save text resources
-            let title = new RowModel(txtResTable!);
-            title.setAttribute({
-                lang: "en",
-                variation: "us",
-                name: colTitleKey,
-                value: colName
-            });
-            title.save(user);
-
-            let description = new RowModel(txtResTable!);
-            description.setAttribute({
-                lang: "en",
-                variation: "us",
-                name: colDescriptionKey,
-                value: "Column Generated by Auria and currently does not have a description"
-            });
-            description.save(user);
-        }
-        catch (err) {
-            console.error("[TableManager] Failed to create column!", err);
-        }*/
+        /*
+                try {
+        
+                    let [colTable, txtResTable] = await Promise.all([
+                        this.module.getTable(user, "Auria.Collumn"),
+                        this.module.getTable(user, "Auria.TextResource")]
+                    );
+        
+                    let colName = this.toCamelCase(comparisson.outOfSync!.field);
+                    let colTitleKey = "@{Auria.Columns.Title." + table.getName() + "." + colName + "}";
+                    let colDescriptionKey = "@{Auria.Columns.Description." + table.getName() + "." + colName + "}";
+                    let dataType = this.rawTypeToDataType(comparisson.outOfSync!.type!);
+        
+                    // # - Create and save column
+                    let col = new RowModel(colTable!);
+                    col.setAttribute({
+                        name: colName,
+                        column: comparisson.outOfSync!.field,
+                        table_name: table.table,
+                        title: colTitleKey,
+                        description: colDescriptionKey,
+                        table_type: "physic",
+                        data_type: dataType,
+                        attributes: comparisson.outOfSync!.key,
+                        required: 1,
+                        allow_modification: 1,
+                        active: 1
+                    });
+                    col.save(user);
+        
+                    // # - Create and save text resources
+                    let title = new RowModel(txtResTable!);
+                    title.setAttribute({
+                        lang: "en",
+                        variation: "us",
+                        name: colTitleKey,
+                        value: colName
+                    });
+                    title.save(user);
+        
+                    let description = new RowModel(txtResTable!);
+                    description.setAttribute({
+                        lang: "en",
+                        variation: "us",
+                        name: colDescriptionKey,
+                        value: "Column Generated by Auria and currently does not have a description"
+                    });
+                    description.save(user);
+                }
+                catch (err) {
+                    console.error("[TableManager] Failed to create column!", err);
+                }*/
     }
 
     public rawTypeToDataType(type: string) {
@@ -221,19 +216,7 @@ export class TableManagerListener extends ModuleListener {
         return { status: "sync" };
     }
 
-    public list: (req: AuriaRequest, res: AuriaResponse) => void =
-        (req, res) => {
-            let tables: any[] = [];
 
-            this.tables.forEach((table) => {
-                tables.push(table.asJSON());
-            });
-
-            res.addToResponse({
-                tables: tables
-            });
-            res.send();
-        };
 
 }
 
